Project: Weather Data Classification using Decision Trees
Objective was to predict afternoon weather by observing the morning weather. I converted this into classification problem by creating the labels. Using the decision tree classifier of SK-Learn library, I created a model to predict the label to forecast afternoon weather. Could also calculate the accuracy of the model by calculating the error of prediction.

This data was collected from a weather station located in San Diego, California. The weather station is equipped with sensors that capture weather-related measurements such as air temperature, air pressure, and relative humidity.  Data was collected for a period of three years, from September 2011 to September 2014, to ensure that sufficient data for different seasons and weather conditions is captured.