Project on Regression Analysis
Objective was to predict the numeric overall rating of a player using linear regressor of SK-Learn package and calculate the accuracy of the prediction using the RMSE (root mean square error) function.

European Soccer Database has more than 25,000 matches and more than 10,000 players for European professional soccer seasons from 2008 to 2016. It is an open dataset available on Kaggle.