Project on Natural Language Processing using NLTK library
Objective was to learn working with NLTK library. As part of this, I did sentiment analysis on movie reviews database available in NLTK corpora. Movie review dataset of NLTK corpora has 2000 files with 800 words on average for each file. 

As part of the project, data in files had to be tokenized; bag of words created after removing stop-words and punctuation, counted term frequencies and using NLTK’s Naive-Bayes-Classifier, converted it into classification problem to predict the sentiment of each text review i.e. if the review is positive or negative.