Sentiment analysis of twitter tweets using twitter API
Goal of this project was to learn analyse the live data generated on twitter by tweets and hash-tag words. Twitter analysis can be very useful in understanding the trends on local and global level. 

As part of the project, I created twitter developer account, accessed the twitter API, fetched the global trends using “Yahoo, where on earth ID (woeid)”, fetched the tweets on the basis of chosen hash-tag word, created the bag of words to understand the top used words and printed frequency distribution of words.