Project_Nearest neighbor for handwritten digit recognition
Purpose of this project was to build a classifier using k-means algorithm which predicts the digit printed on the image in a hand-written style. We used MNIST dataset in this project. MNIST dataset is the collection of 60000 images (28x28 pixels) of handwritten digits, each labelled with its correct digit. It has another 10000 images and their labels as test set. 

Approach was to create a function which will give the index of the vector of minimum euclidean distance from the given vector out of all the vectors. Thus, we will get the label of the predicted vector. We also compared the performance of our classifier with in-built classifiers of sklearn.neighbors library such as BallTree and KDTree