Project_Winery classification using the one & multi-dimensional Gaussian
Wine dataset is one of the dataset from UCI repository. The UCI Machine Learning Repository is a collection of databases, domain theories, and data generators that are used by the machine learning community for the empirical analysis of machine learning algorithms.

Wine data is the result of a chemical analysis of wines grown in the same region in Italy but derived from three different cultivars. The analysis determined the quantities of 13 constituents found in each of the three types of wines. It contains 178 labelled data points, each corresponding to a type of wine. 

As part of this assignment, purpose was to build a classifier using the Gaussian generative modelling approach to predict the type of wine using only one feature (uni-variate), two features (bi-variate) and all the 13 features (multi-variate).