Data Cleaning Project
In this project, we will use different txt and CSV files and clean them to ready for data analysis. As part of this project, we will drop unnecessary columns, take care of null values, change the index of the dataframe, observe the data types of each column, prepare them to convert into relevant pandas data types to make it ready for further analysis.

We will also use applymap() function with custom function to clean different coulumns in one go. We will also perform renaming of columns & lables and skipping rows.