Classification problem with one hidden layer Neural Networ
Purpose of this project is build a classifier to define regions as either red or blue using neural network with one hidden layer for this classification problem. I will also compare it with logistic regression model to compare the peformance. We will also assess our model on different matrices for classification problems.